Move new config defs to already exported package, and add to CMakeLists
