+++ container-di/src/main/resources/configdefinitions/platform-bundles.def
@@ -1,5 +1,5 @@
 # Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package=com.yahoo.container.di
+package=com.yahoo.container.di.config
 
 # References to platform bundles to install.
 bundles[] file
